removed simple head and hqsam head as options
