+++ kestrel/config.py
@@ -18,9 +18,6 @@ class ModelPaths:
     reference_weights: Optional[Path] = None
     reference_config: Optional[Path] = None
     reference_tokenizer: Optional[str | Path] = None
-    head_refiner_weights: Optional[Path] = None
-    sam_head_refiner_weights: Optional[Path] = None
-    hqsam_head_refiner_weights: Optional[Path] = None
 
 
 @dataclass
@@ -34,7 +31,6 @@ class RuntimeConfig:
     page_size: int = 128
     max_seq_length: int = 32768
     enable_cuda_graphs: bool = True
-    refiner_type: str = "head"  # "head", "sam", "samhead", or "hqsamhead"
     refiner_iters: int = 6
 
     def resolved_dtype(self) -> torch.dtype:
