+++ seg_repl.py
@@ -1,32 +1,14 @@
 """
 uv run seg_repl.py \
   --kestrel-weights /ephemeral/vixtral-train/ckpt/rl/seg-warp-3_1/s941/model.pt \
-  --refiner-weights /ephemeral/vixtral-train/ckpt/mask-refiner/test_train_refiner_01/s31000/model.pt \
-  --refiner head \
   --refiner-iters 5
-
-# Or with samhead:
-# uv run seg_repl.py \
-#   --kestrel-weights /ephemeral/vixtral-train/ckpt/rl/seg-warp-3_1/s941/model.pt \
-#   --refiner-weights /path/to/samhead/model.pt \
-#   --refiner samhead \
-#   --refiner-iters 5
-
-# Or with hqsamhead:
-# uv run seg_repl.py \
-#   --kestrel-weights /ephemeral/vixtral-train/ckpt/rl/seg-warp-3_1/s941/model.pt \
-#   --refiner-weights /path/to/hqsamhead/model.pt \
-#   --refiner hqsamhead \
-#   --refiner-iters 5
 """
 
 import argparse
 import asyncio
 import os
-import sys
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Optional
 
 import cv2
 import numpy as np
@@ -35,14 +17,12 @@ from PIL import Image
 
 from kestrel.config import ModelPaths, RuntimeConfig
 from kestrel.engine import InferenceEngine
-from kestrel.seg_refiner import svg_from_path, render_svg_to_soft_mask
+from kestrel.head_seg_refiner import svg_from_path, render_svg_to_soft_mask
 
 
 @dataclass
 class REPLConfig:
     kestrel_weights: Path
-    refiner_weights: Optional[Path]
-    refiner: str
     refiner_iters: int
     device: str
     temperature: float
@@ -105,15 +85,11 @@ async def init_engine(config: REPLConfig) -> InferenceEngine:
     runtime_cfg = RuntimeConfig(
         model_paths=ModelPaths(
             weights=config.kestrel_weights,
-            head_refiner_weights=config.refiner_weights if config.refiner == "head" else None,
-            sam_head_refiner_weights=config.refiner_weights if config.refiner == "samhead" else None,
-            hqsam_head_refiner_weights=config.refiner_weights if config.refiner == "hqsamhead" else None,
         ),
         device=config.device,
         dtype=torch.bfloat16,
         max_batch_size=2,
         enable_cuda_graphs=False,
-        refiner_type=config.refiner,
         refiner_iters=config.refiner_iters,
     )
 
@@ -123,7 +99,7 @@ async def init_engine(config: REPLConfig) -> InferenceEngine:
 
 async def repl_loop(engine: InferenceEngine, config: REPLConfig):
     print(f"\nKestrel Segmentation REPL")
-    print(f"Refiner: {config.refiner} (iters={config.refiner_iters})")
+    print(f"Refiner iters: {config.refiner_iters}")
     print(f"Temperature: {config.temperature}")
     print(f"Press Ctrl+C to exit\n")
 
@@ -142,7 +118,7 @@ async def repl_loop(engine: InferenceEngine, config: REPLConfig):
             continue
 
         base = os.path.splitext(os.path.basename(image_path))[0]
-        output_path = f"{base}_seg_{config.refiner}_{iteration}.png"
+        output_path = f"{base}_seg_{iteration}.png"
 
         print(f"\nProcessing: {image_path}")
         print(f"Prompt: {text_prompt}")
@@ -159,20 +135,13 @@ async def repl_loop(engine: InferenceEngine, config: REPLConfig):
 async def async_main():
     parser = argparse.ArgumentParser()
     parser.add_argument("--kestrel-weights", required=True, type=Path)
-    parser.add_argument("--refiner-weights", type=Path)
-    parser.add_argument("--refiner", choices=["head", "sam", "samhead", "hqsamhead"], required=True)
     parser.add_argument("--refiner-iters", type=int, default=6)
     parser.add_argument("--temperature", type=float, default=0.0)
     parser.add_argument("--device", default="cuda")
     args = parser.parse_args()
 
-    if args.refiner in ("head", "samhead", "hqsamhead") and not args.refiner_weights:
-        sys.exit(f"--refiner-weights required with --refiner {args.refiner}")
-
     config = REPLConfig(
         kestrel_weights=args.kestrel_weights,
-        refiner_weights=args.refiner_weights,
-        refiner=args.refiner,
         refiner_iters=args.refiner_iters,
         device=args.device,
         temperature=args.temperature,
